codecrafters submit [skip ci]

+++ src/main.c
@@ -48,7 +48,7 @@ int main() {
     return 1;
   }
 
-  int bytesRead;
+  ssize_t bytesRead;
   char buffer[512];
   socklen_t clientAddrLen = sizeof(clientAddress);
 
@@ -62,7 +62,7 @@ int main() {
     }
 
     buffer[bytesRead] = '\0';
-    printf("Received %d bytes: %s\n", bytesRead, buffer);
+    printf("Received %zd bytes: %s\n", bytesRead, buffer);
 
     struct flags f = {htons(1), 0, 0, 0, 0, 0, 0, htons(1)};
     DNSHeader header = dns_header_new(1234, f, 0, 0, 0, 0);
